backend/pyproject.toml
[tool.poetry]
name = "greensentinel-backend"
version = "0.1.0"
description = "GreenSentinel API backend"
authors = ["GreenSentinel Team"]
readme = "README-backend.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
bcrypt = "^4.1.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
geoalchemy2 = "^0.14.5"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
minio = "^7.2.5"
python-multipart = "^0.0.9"
openai = "^1.30.0"
aio_pika = "^9.3.1"
pydantic-msgspec = "^0.5.4"
fastapi-websocket-pubsub = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.5"
ruff = "^0.4.5"
respx = "^0.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "UP", "ANN", "B"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
